+++ app/Jobs/SendNewsletterEmail.php
@@ -219,6 +219,9 @@ class SendNewsletterEmail implements ShouldQueue
                     ->success()
                     ->sendToDatabase($user);
             }
+
+            $message->refresh();
+            $message->purgeSendsForTestingAudience();
         }
     }
 }